Update tests for PositionsMerkleTree
